refactor frontend job package
