+++ frontend/src/components/Routes.jsx
@@ -14,7 +14,9 @@ const Organization = React.lazy(() => import("./me/organization/Organization"));
 const Sessions = React.lazy(() => import("./me/sessions/Sessions"));
 const JobsTable = React.lazy(() => import("./jobs/table/JobsTable"));
 const JobResult = React.lazy(() => import("./jobs/result/JobResult"));
-const CommentResult = React.lazy(() => import("./jobs/result/CommentResult"));
+const CommentResult = React.lazy(() =>
+  import("./jobs/result/bar/comment/CommentResult"),
+);
 const PluginsContainer = React.lazy(() => import("./plugins/PluginsContainer"));
 const Dashboard = React.lazy(() => import("./dashboard/Dashboard"));
 const ScanForm = React.lazy(() => import("./scan/ScanForm"));
