--- frontend/src/components/jobs/result/utils/index.js
@@ -1,2 +1,0 @@
-export { default as JobOverview } from "./JobOverview";
-export { default as CommentOverview } from "./CommentOverview";
