+++ frontend/src/utils/time.js
@@ -1,0 +1,3 @@
+export function processTimeMMSS(value) {
+  return new Date(value * 1000).toISOString().substring(14, 19);
+}
